@@ -3,13 +3,11 @@ namespace MVFC.Aspire.Helpers.Mongo.Factories;
 /// <summary>
 /// Implementação da factory de clientes MongoDB
 /// </summary>
-internal sealed class MongoClientFactory : IMongoClientFactory
+internal static class MongoClientFactory
 {
-    private readonly IMongoDumpProcessor _dumpProcessor = new MongoDumpProcessor();
+    private static readonly ConcurrentDictionary<string, IMongoClient> _clients = new();
 
-    private static readonly System.Collections.Concurrent.ConcurrentDictionary<string, IMongoClient> _clients = new();
-
-    public async Task ExecuteDumpsAsync(
+    internal static async Task ExecuteDumpsAsync(
         string connectionString,
         IReadOnlyCollection<IMongoClassDump> dumps,
         CancellationToken cancellationToken)
@@ -17,14 +15,9 @@ internal sealed class MongoClientFactory : IMongoClientFactory
         ArgumentException.ThrowIfNullOrWhiteSpace(connectionString);
         ArgumentNullException.ThrowIfNull(dumps);
 
-        if (dumps.Count == 0)
-        {
-            return;
-        }
-
         var client = _clients.GetOrAdd(connectionString, CreateClient);
 
-        await _dumpProcessor.ProcessDumpsAsync(client, dumps, cancellationToken);
+        await MongoDumpProcessor.ProcessDumpsAsync(client, dumps, cancellationToken);
     }
 
     private static IMongoClient CreateClient(string connectionString)
