@@ -1,6 +1,5 @@
-﻿global using Bogus;
+global using Bogus;
 global using System.Reflection;
-global using System.Diagnostics.CodeAnalysis;
 global using Aspire.Hosting;
 global using Aspire.Hosting.ApplicationModel;
 global using MongoDB.Driver;
@@ -9,4 +8,4 @@ global using MVFC.Aspire.Helpers.Mongo.Models;
 global using MVFC.Aspire.Helpers.Mongo.Factories;
 global using MVFC.Aspire.Helpers.Mongo.Processors;
 global using MVFC.Aspire.Helpers.Mongo.Helpers;
-global using MVFC.Aspire.Helpers.Mongo.Extensions;
+global using System.Collections.Concurrent;
